requirements.txt
absl-py==0.12.0
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
audioread==2.1.9
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
Bottleneck==1.3.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
ConfigArgParse==1.4
configparser==5.0.2
ctc-segmentation==1.5.2
cycler==0.10.0
Cython==0.29.23
decorator==5.0.7
defusedxml==0.7.1
Distance==0.1.3
docker-pycreds==0.4.0
editdistance==0.5.2
entrypoints==0.3
espnet==0.9.9
espnet-model-zoo==0.0.0a29
espnet-tts-frontend==0.0.3
filelock==3.0.12
flatbuffers==1.12
g2p-en==2.1.0
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
humanfriendly==9.1
idna==2.10
inflect==5.3.0
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jaconv==0.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaldiio==2.17.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.1.0
matplotlib-inline==0.1.2
mistune==0.8.4
nara-wpe==0.0.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.3.0
numba==0.53.1
numpy==1.20.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkg-resources==0.0.0
pooch==1.3.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.16.0
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pypinyin==0.41.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-wpe==0.0.1
pytz==2021.1
pyworld==0.2.12
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
sentencepiece==0.1.86
sentry-sdk==1.1.0
shortuuid==1.0.1
six==1.15.0
smmap==4.0.0
SoundFile==0.10.3.post1
streamlit==0.81.1
subprocess32==3.5.4
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
torch==1.8.1
torch-complex==0.2.1
torchaudio==0.8.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typeguard==2.12.0
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.26.4
validators==0.18.2
wandb==0.10.30
watchdog==2.1.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
